// src/generated/types/CreateDeviceDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { DeviceSpecificationDto } from './DeviceSpecificationDto.ts'

export type CreateDeviceDto = {
  /**
   * @type string
   */
  IPAddress: string
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  labId: string
  /**
   * @type string
   */
  assisstantId: string
  /**
   * @type array | undefined
   */
  specifications?: DeviceSpecificationDto[] | undefined
}