// src/generated/types/UpdateDeviceDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { DeviceSpecificationDto } from './DeviceSpecificationDto.ts'

export type UpdateDeviceDto = {
  /**
   * @type string | undefined
   */
  IPAddress?: string | undefined
  /**
   * @type string | undefined
   */
  name?: string | undefined
  /**
   * @type string | undefined
   */
  labId?: string | undefined
  /**
   * @type string | undefined
   */
  assisstantId?: string | undefined
  /**
   * @type array | undefined
   */
  specifications?: DeviceSpecificationDto[] | undefined
}